Rename 'i' to numClientCodes1  (Silly oversight here)
The variable i had an important meaning as numClientCodes1 in the beginning
But I totally missed it was also being used as a regular counter later on, and
also replaced those usages. It was at the very end of the code anyway.

I come back to fix it later in the evening when I look down at the last section
of the code.