First performance fix attempt
Improves the look of the code, but unsuccessful.
I change loops to a one-step copy/paste.

I added an extra performance counter to try to locate hotpaths.

This narrowed it down to the loops I mentioned the O(N^2) about.
However it turned out to be unrelated to the complexity, and instead
having to do with those loops copying cells through the clipboard.